+++ PCbuild/pythoncore.vcxproj.filters
@@ -60,6 +60,9 @@
     <ClInclude Include="..\Include\bytesobject.h">
       <Filter>Include</Filter>
     </ClInclude>
+    <ClInclude Include="..\Include\call.h">
+      <Filter>Include</Filter>
+    </ClInclude>
     <ClInclude Include="..\Include\cellobject.h">
       <Filter>Include</Filter>
     </ClInclude>
